getDistance accept only one argument - Point теперь можно вводить свои координаты с html думаю на координатами для  Многоугольника
