+++ HW10/Index.html
@@ -17,13 +17,14 @@
       <li> багатокутник:-> прямокутник -> довжина, висота, площа(), периметр()</li>
       <li> прямокутник: -> квадрат -> довжинаСторони , площа(), периметр()</li>
     </ul>
-    <!--<p class="error"></p>-->
+    <p class="error"></p>
 
-    <!--<div class="form">-->
-      <!--<input class="txt js_x" placeholder=" Введите x">-->
-      <!--<input class="txt js_y" placeholder=" Введите У">-->
-      <!--<button class="btn js_bnt">Enter</button>-->
-    <!--</div>-->
+    <div class="form">
+      <input class="txt js_x" placeholder=" Введите x">
+      <input class="txt js_y" placeholder=" Введите У">
+      <input class="txt js_points" placeholder=" Введите point">
+      <button class="btn js_bnt">Enter</button>
+    </div>
   </div>
   <!--<div class="popup hide">-->
     <!--<header class="popup_header">-->
